Resolved issue with combine reducers
